+++ index.html
@@ -32,5 +32,11 @@
     </div>
   </body>
 
-  <script src="script.js"></script>
+  <script type="module" src="script.js"></script>
+  <script type="module" src="render.js"></script>
+  <script type="module" src="like-button.js"></script>
+  <script type="module" src="add-reply.js"></script>
+  <script type="module" src="api.js"></script>
+  <script type="module" src="format-date.js"></script>
+
 </html>
